rename-config-file - renamed config.js to Config.js - updated references to config.js
